api/requirements.txt
fastapi
uvicorn
jinja2
pydantic
openai
python-multipart
python-dotenv
